LPS-95015 Move test method to module. LPS-72818 (53e82f47e2586cb4d940374b7b218d2a730d2d9d) moved the test to site-test module. Later, LPS-85085 (e6b933ee3fe4514a19e2a7b91b1a34228a9ed17d) added the test to portal-impl. So they should be same test class.
