+++ modules/apps/site/site-test/src/testIntegration/java/com/liferay/site/service/persistence/test/GroupServiceTest.java
@@ -61,6 +61,7 @@ import com.liferay.portal.kernel.util.FriendlyURLNormalizerUtil;
 import com.liferay.portal.kernel.util.LocaleUtil;
 import com.liferay.portal.kernel.util.PortalUtil;
 import com.liferay.portal.kernel.util.PortletKeys;
+import com.liferay.portal.kernel.util.StringUtil;
 import com.liferay.portal.test.rule.Inject;
 import com.liferay.portal.test.rule.LiferayIntegrationTestRule;
 import com.liferay.portal.test.rule.PermissionCheckerMethodTestRule;
@@ -85,6 +86,7 @@ import org.junit.Test;
 import org.junit.runner.RunWith;
 
 /**
+ * @author Drew Brokke
  * @author Julio Camarero
  * @author Roberto Díaz
  * @author Sergio González
@@ -597,6 +599,101 @@ public class GroupServiceTest {
 			defaultUser.setLanguageId(languageId);
 
 			UserLocalServiceUtil.updateUser(defaultUser);
+		}
+	}
+
+	@Test
+	public void testGetGroupsLikeName() throws Exception {
+		List<Group> allChildGroups = new ArrayList<>();
+		Group parentGroup = GroupTestUtil.addGroup();
+
+		List<Group> allGroups = new ArrayList<>(
+			GroupLocalServiceUtil.getGroups(
+				TestPropsValues.getCompanyId(),
+				GroupConstants.DEFAULT_PARENT_GROUP_ID, true));
+
+		try {
+			String name = RandomTestUtil.randomString(10);
+
+			long parentGroupId = parentGroup.getGroupId();
+
+			List<Group> likeNameChildGroups = new ArrayList<>();
+
+			for (int i = 0; i < 10; i++) {
+				Group group = GroupTestUtil.addGroup(parentGroupId);
+
+				group.setName(name + i);
+
+				group = GroupLocalServiceUtil.updateGroup(group);
+
+				likeNameChildGroups.add(group);
+			}
+
+			allChildGroups.addAll(likeNameChildGroups);
+			allChildGroups.add(GroupTestUtil.addGroup(parentGroupId));
+			allChildGroups.add(GroupTestUtil.addGroup(parentGroupId));
+			allChildGroups.add(GroupTestUtil.addGroup(parentGroupId));
+
+			allGroups.addAll(allChildGroups);
+
+			assertExpectedGroups(
+				likeNameChildGroups, parentGroupId, name + "%");
+			assertExpectedGroups(
+				likeNameChildGroups, parentGroupId,
+				StringUtil.toLowerCase(name) + "%");
+			assertExpectedGroups(
+				likeNameChildGroups, parentGroupId,
+				StringUtil.toUpperCase(name) + "%");
+			assertExpectedGroups(
+				likeNameChildGroups, GroupConstants.ANY_PARENT_GROUP_ID,
+				name + "%");
+			assertExpectedGroups(allChildGroups, parentGroupId, null);
+			assertExpectedGroups(allChildGroups, parentGroupId, "");
+			assertExpectedGroups(
+				allGroups, GroupConstants.ANY_PARENT_GROUP_ID, "");
+		}
+		finally {
+			for (Group childGroup : allChildGroups) {
+				GroupTestUtil.deleteGroup(childGroup);
+			}
+
+			GroupTestUtil.deleteGroup(parentGroup);
+		}
+	}
+
+	@Test
+	public void testGetGtGroups() throws Exception {
+		for (int i = 0; i < 10; i++) {
+			_groups.add(GroupTestUtil.addGroup());
+		}
+
+		long parentGroupId = 0;
+		int size = 5;
+
+		List<Group> groups = GroupServiceUtil.getGtGroups(
+			0, TestPropsValues.getCompanyId(), parentGroupId, true, size);
+
+		Assert.assertFalse(groups.isEmpty());
+		Assert.assertEquals(groups.toString(), size, groups.size());
+
+		Group lastGroup = groups.get(groups.size() - 1);
+
+		groups = GroupServiceUtil.getGtGroups(
+			lastGroup.getGroupId(), TestPropsValues.getCompanyId(),
+			parentGroupId, true, size);
+
+		Assert.assertFalse(groups.isEmpty());
+		Assert.assertEquals(groups.toString(), size, groups.size());
+
+		long previousGroupId = 0;
+
+		for (Group group : groups) {
+			long groupId = group.getGroupId();
+
+			Assert.assertTrue(groupId > lastGroup.getGroupId());
+			Assert.assertTrue(groupId > previousGroupId);
+
+			previousGroupId = groupId;
 		}
 	}
 
@@ -1203,6 +1300,27 @@ public class GroupServiceTest {
 			null);
 	}
 
+	protected void assertExpectedGroups(
+			List<Group> expectedGroups, long parentGroupId, String nameSearch)
+		throws Exception {
+
+		List<Group> actualGroups = GroupServiceUtil.getGroups(
+			TestPropsValues.getCompanyId(), parentGroupId, nameSearch, true,
+			QueryUtil.ALL_POS, QueryUtil.ALL_POS);
+
+		Assert.assertEquals(
+			actualGroups.toString(), expectedGroups.size(),
+			actualGroups.size());
+		Assert.assertTrue(
+			actualGroups.toString(), actualGroups.containsAll(expectedGroups));
+
+		Assert.assertEquals(
+			expectedGroups.size(),
+			GroupServiceUtil.getGroupsCount(
+				TestPropsValues.getCompanyId(), parentGroupId, nameSearch,
+				true));
+	}
+
 	protected Locale getLocale() {
 		ThemeDisplay themeDisplay = new ThemeDisplay();
 
@@ -1290,6 +1408,9 @@ public class GroupServiceTest {
 
 	@DeleteAfterTestRun
 	private Group _group;
+
+	@DeleteAfterTestRun
+	private final List<Group> _groups = new ArrayList<>();
 
 	@Inject
 	private PortletPreferencesLocalService _portletPreferencesLocalService;
